Use same initialization in batch (#5)
* Use same initialization in batch

In SPSA, randomization was done for each input in a batch. This made the
implementation batch inconsistent.

* Stop expanding x and y repeatedly

In SPSA, the expansion of x and y are done multiple times in a for loop.
This operation is done once outside the loop now.

* Truncate v_ correctly

The directions of perturbation in spspa is stored in v_. This needs to
be truncated if the last batch is small. The case of small last batch is
also tested.